% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertAssignmentPipe.R
\name{insertAssignmentPipe}
\alias{insertAssignmentPipe}
\title{Insert \verb{\%<>\%}}
\usage{
insertAssignmentPipe()
}
\description{
Inserts \code{\link[magrittr:compound]{\%<>\%}} at the cursor position.
}
