% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertExpositionPipe.R
\name{insertExpositionPipe}
\alias{insertExpositionPipe}
\title{Insert \verb{\%$\%}}
\usage{
insertExpositionPipe()
}
\description{
Inserts \code{\link[magrittr:exposition]{\%$\%}} at the cursor position.
}
