% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMagrittrCompoundAssignment.R
\name{insertMagrittrCompoundAssignment}
\alias{insertMagrittrCompoundAssignment}
\title{Inserts \%<>\%}
\usage{
insertMagrittrCompoundAssignment()
}
\description{
Insert \code{\%\<\>\%}
}
