% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMagrittrExpositionOp.R
\name{insertMagrittrExpositionOperator}
\alias{insertMagrittrExpositionOperator}
\title{Insert \verb{\%$\%}}
\usage{
insertMagrittrExpositionOperator()
}
\description{
Inserts \code{\link[magrittr:exposition]{\%$\%}} at the cursor position.
}
