% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertTeeMagrittrOp.R
\name{insertMagrittrTeeOperator}
\alias{insertMagrittrTeeOperator}
\title{Insert \verb{\%T>\%}}
\usage{
insertMagrittrTeeOperator()
}
\description{
Inserts \code{\link[magrittr:tee]{\%T>\%}} at the cursor position.
}
