% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertNULLDefault.R
\name{insertNULLDefault}
\alias{insertNULLDefault}
\title{Insert \verb{\%||\%}}
\usage{
insertNULLDefault()
}
\description{
Inserts \code{\link[rlang:op-null-default]{\%||\%}} at the cursor position.
}
