% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertRightAssignment.R
\name{insertRightAssignment}
\alias{insertRightAssignment}
\title{Insert \verb{->}}
\usage{
insertRightAssignment()
}
\description{
Inserts \code{\link[base:assignOps]{->}} at the cursor position.
}
