% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utility_functions}
\alias{utility_functions}
\alias{nextToSpace}
\title{Generic utility functions}
\usage{
nextToSpace()
}
\description{
Generic utility functions usually not to be used outside the package
environment.
}
\section{Check space}{

Check if last character is an empty space.
}

\keyword{internal}
