% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMagrittrCompoundAssignment.R
\name{insertMagrittrCompoundAssignment}
\alias{insertMagrittrCompoundAssignment}
\title{Insert \verb{\%<>\%}}
\usage{
insertMagrittrCompoundAssignment()
}
\description{
Inserts \code{\link[magrittr:compound]{\%<>\%}} at the cursor position.
}
